RFS-170 - Changed devnet config to respect same structure as the rest of the networks.
