+++ rskj-core/src/main/java/org/ethereum/config/blockchain/devnet/DevNetFirstForkConfig.java
@@ -17,7 +17,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.ethereum.config.blockchain;
+package org.ethereum.config.blockchain.devnet;
 
 import co.rsk.config.BridgeConstants;
 import co.rsk.config.BridgeDevNetConstants;
@@ -26,28 +26,9 @@ import org.ethereum.config.blockchain.testnet.TestNetAfterBridgeSyncConfig;
 /**
  * Created by Oscar Guindzberg on 25.10.2016.
  */
-public class DevNetConfig extends TestNetAfterBridgeSyncConfig {
+public class DevNetFirstForkConfig extends DevNetGenesisConfig {
 
-    public static class DevNetConstants extends TestNetConstants {
-        private static final byte CHAIN_ID = 32;
-        @Override
-        public BridgeConstants getBridgeConstants() {
-            return BridgeDevNetConstants.getInstance();
-        }
-
-        @Override
-        public byte getChainId() {
-            return DevNetConstants.CHAIN_ID;
-        }
-    }
-
-    public DevNetConfig() {
-        super(new DevNetConstants());
-    }
-
-    @Override
-    public boolean areBridgeTxsFree() {
-        return true;
+    public DevNetFirstForkConfig() {
     }
 
     @Override
